API Reference

windc_household_data.clean_cps_data_year — Method

clean_cps_data_year(year, cps_rw, variables, api_key; states = STATES)

Download and clean the CPS data for a given year.

windc_household_data.cps_vs_nipa_income_categories — Method

cps_vs_nipa_income_categories(income::DataFrame, api_key, years)

Return a comparison between CPS and NIPA income categories.

Arguments

income - A DataFrame with the CPS income data

api_key - Request an API key from the BEA website

years - The years to pull data for, as a range 2000:2023

Returns

A DataFrame with the columns year, nipa, cps, pct_diff, and category.

windc_household_data.get_bea_nipa_data — Method

get_bea_nipa_data(bea_key::String, years)

Get NIPA data from the BEA API for the given years.

Arguments

bea_key - Request an API key from the BEA website

years - The years to pull data for, as a range 2001:2024 -> These are +1 for some reason

Returns

A DataFrame with the NIPA data for the given years and columns year, LineNumber, and nipa.

API query

query = Dict(
    "UserID" => bea_key,
    "method" => "GetData",
    "datasetname" => "NIPA",
    "TableName" => "T20100",
    "Frequency" => "A",
    "Year" => join(years, ","),
    "ResultFormat" => "JSON"
)

windc_household_data.get_cps_data_api — Method

get_cps_data_api(year, vars, api_key)

Pull the raw data directly from the API

windc_household_data.load_cps_data_api — Method

load_cps_data_api(api_key, years; states = STATES)

Load all the CPS data in the given year range.

Arguments

api_key - Request an API key from the Census website

years - The years to pull data for, as a range 2000:2023

Keywords

states - A DataFrame with the state FIPS codes and abbreviations (default is STATES)

Returns

A dictionary with the following keys:

income - A DataFrame with the income data

shares - A DataFrame with the income shares

count - A DataFrame with the count of households

numhh - A DataFrame with the number of households

API Call

There is a split in 2019.

Common	Pre-2019	Post-2019	Description

HWSVAL			wages and salaries

HSEVAL			self-employment (nonfarm)

HFRVAL			self-employment farm

HUCVAL			unemployment compensation

HWCVAL			workers compensation

HSSVAL			social security

HSSIVAL			supplemental security

HPAWVAL			public assistance or welfare

HVETVAL			veterans benefits

HSURVAL			survivors income

HDISVAL			disability

HINTVAL			interest

HDIVVAL			dividends

HRNTVAL			rents

HEDVAL			educational assistance

HCSPVAL			child support

HFINVAL			financial assistance

HOIVAL			other income

HTOTVAL			total household income

GESTFIPS		state fips

A_EXPRRP		expanded relationship code

H_HHTYPE		type of household interview

PPPOS		person identifier

MARSUPWT		asec supplement final weight

HDSTVAL	retirement distributions

HPENVAL	pension income

HANNVAL	annuities