Introduction

This package contains method to download the raw CSV data used in the WiNDC.GAMS build stream.

To add this package to your Julia environment, open the package manager and type:

pkg> add https://github.com/uw-windc/windc_household_data

Download the data is a single function call:

using windc_household_data

census_api_key = "<census_api_key>"
bea_api_key = "<bea_api_key>"

years = 2000:2023

download_save_data(
    census_api_key, 
    bea_api_key, 
    years, 
    "data/";
    acs_year = 2020
    )

API Reference

windc_household_data.acs_commuting — Method

acs_commuting(raw_commuting::DataFrame, wages::DataFrame)

Compute the commuting data from the ACS data and the wages data.

Arguments

raw_commuting - The raw commuting data from the ACS, output from read_commuting_data.

wages - The wages data from the CPS data, output from cps_wages.

windc_household_data.clean_cps_data_year — Method

clean_cps_data_year(year, cps_rw, variables, api_key; states = STATES)

Download and clean the CPS data for a given year.

windc_household_data.cps_count_observations — Method

cps_count_observations(cps::DataFrame)

Count the number of observations in the CPS data by household and state.

windc_household_data.cps_income — Method

cps_income(cps::DataFrame)

Compute the income from the CPS data by variable.

windc_household_data.cps_income_shares — Method

cps_income_shares(cps::DataFrame)

Compute the income share for by household.

windc_household_data.cps_number_households — Method

cps_number_households(cps::DataFrame; unit::Real = 1e-6)

Compute the number of households from the CPS data. This is the sum of the variable marsupwt for each household and state.

windc_household_data.cps_vs_nipa_income_categories — Method

cps_vs_nipa_income_categories(income::DataFrame, api_key::String)

Return a comparison between CPS and NIPA income categories.

Arguments

income - A DataFrame with the CPS income data

api_key - Request an API key from the BEA website

Returns

A DataFrame with the columns year, nipa, cps, pct_diff, and category.

windc_household_data.cps_wages — Method

cps_wages(census_data::DataFrame; year = 2020)

Compute the wages from the CPS data by state and year.

Arguments

census_data - The CPS data from the API or a CSV file, output from load_cps_data_api.

year - The year to pull the data for. Default is 2020. The commuting data is released every 5 years, this is the most recent year or release.

windc_household_data.create_nipa_windc — Method

create_nipa_windc(bea_data::DataFrame, type::Symbol)

Create a DataFrame comparing the WiNDC vs NIPA data. This uses two magic files, ld0_windc.csv and kd0_windc.csv to create the comparison.

Arguments

bea_data - A DataFrame with the NIPA data

type - Either :labor or :capital

windc_household_data.download_save_data — Method

function download_save_data(
    census_api_key, 
    bea_api_key, 
    years,
    output_root_directory; 
    states = STATES,
    acs_file_name = "table1.xlsx",
    acs_year = 2020
)

Download and save the raw household data and save it into the correct directory structure.

Arguments

census_api_key - The Census API key. Request an API key from the Census website

bea_api_key - The BEA API key. Request an API key from the BEA website

years - The years to pull data for, as a range 2000:2023

output_root_directory - The directory to save the data

Optional Arguments

states - A DataFrame with the state FIPS and abbreviations. Default is STATES.

acs_file_name - The name of the ACS commuting data file. Default is "table1.xlsx".

acs_year - The year of the ACS data. Default is 2020.

Functions Called

load_cps_data_api

get_bea_nipa_data

save_cps_data

magic_data_cps

get_cms_data

extrapolate_cms_data

get_census_health_data

public_health_benefits

save_public_health_benefits_data

get_acs_commuting_data

save_acs_commuting_data

windc_household_data.extrapolate_cms_data — Method

extrapolate_cms_data(cms::DataFrame, years)

Extrapolate CMS data for each given year. Performs the extrapolation for each state independently.

windc_household_data.extrapolate_cms_data_single_state — Method

extrapolate_cms_data_single_state(df, years)

Extrapolate CMS data to years that are not in the data. This uses a linear model to extrapolate the data. Will filter years to be only years not present in data.

Arguments

df::DataFrame: The CMS data.

years::Vector{Int}: All years needed in final data

Returns

A DataFrame with the following columns:

:state::String: The state name.

:year::Int: The year of the data.

:medicare::Float64: The medicare spending.

:medicade::Float64: The medicade spending.

windc_household_data.fetch_commuting_data — Method

fetch_commuting_data(output_directory; year = 2020, file_name = "table1.xlsx", output_name = file_name)

Download the commuting data from the Census Bureau website.

Arguments

output_directory - The directory to save the data.

Optional Arguments

year - The year to pull the data for. Default is 2020. The commuting data is released every 5 years, this is the most recent year or release.

file_name - The name of the file to download. Default is "table1.xlsx".

output_name - The name of the file to save the data as. Default is the same as file_name.

Returns

The path to the downloaded file.

windc_household_data.fetch_zip_data — Method

fetch_zip_data(
    url::String,
    filter_function::Function;;
    output_path::String = joinpath(pwd(), "data"),
)

Download a zip file from a given url and extract the files in the zip file that are in the data NamedTuple.

This function will throw an error if not all files in data are extracted.

Required Arguments

url::String: The url of the zip file to download.

filter_function::Function;: A function that takes a string and returns a boolean. This function is used to filter the files in the zip file, it should return true if the file should be extracted and false otherwise.

Optional Arguments

output_path::String: The path to save the extracted files. Default is the

directory data in the current working directory. If this is not an absolute path, it will be joined with the current working directory.

Output

Returns a vector of the absolute paths to the extracted files.

windc_household_data.get_acs_commuting_data — Method

get_acs_commuting_data(census_data; year = 2020, file_name = "table1.xlsx")

Download the ACS data and compute the commuting data.

Arguments

census_data - The CPS data from the API or a CSV file, output from load_cps_data_api.

Optional Arguments

year - The year to pull the data for. Default is 2020. The commuting data is released every 5 years, this is the most recent year or release.

file_name - The name of the file to download. Default is "table1.xlsx".

windc_household_data.get_bea_nipa_data — Method

get_bea_nipa_data(bea_key::String, years)

Get NIPA data from the BEA API for the given years.

Arguments

bea_key - Request an API key from the BEA website

years - The years to pull data for, as a range 2001:2024 -> These are +1 for some reason

Returns

A DataFrame with the NIPA data for the given years and columns year, LineNumber, and nipa.

API query

query = Dict(
    "UserID" => bea_key,
    "method" => "GetData",
    "datasetname" => "NIPA",
    "TableName" => "T20100",
    "Frequency" => "A",
    "Year" => join(years, ","),
    "ResultFormat" => "JSON"
)

windc_household_data.get_census_health_data — Method

get_census_health_data(census_api_key, years)

Get health data from the ACS for multiple years. This function will return the data for all states.

Retrieves data from the following variables in the ACS 1 year survey:

B27015_005
B27015_010
B27015_015
B27015_020
B27015_025

No data is available for 2020 due to low response rates. We will use the average of 2019 and 2021 data for 2020.

Arguments

census_api_key::String: The Census API key.

years::Vector{Int}: The years to get the data, will filter to be >= 2009.

windc_household_data.get_census_health_data_year — Method

get_census_health_data_year(census_api_key, variables, year)

Get health data from the ACS for a single year. This function will return the data for all states.

We only want estimated data, not MOE data. Will only return estimates.

windc_household_data.get_cms_data — Method

get_cms_data(output;files_to_grab = (medicare = "MEDICARE_AGGREGATE20.CSV", medicaid = "MEDICAID_AGGREGATE20.CSV"))

Download and extract CMS a state health expediture by state of residence data. This downloads a ZIP file from the CMS website.

Arguments

output::String: The directory to save the data to.

Keyword Arguments

files_to_grab::Tuple: A tuple of file names to extract from the ZIP file. The keys are the names of the files and the values are the names of the files in the ZIP file.

Returns

A DataFrame with the following columns:

:state::String: The state name.

:year::Int: The year of the data.

:medicare::Float64: The medicare spending.

:medicade::Float64: The medicaid spending.

windc_household_data.get_cps_data_api — Method

get_cps_data_api(year, vars, api_key)

Pull the raw data directly from the API

windc_household_data.load_cps_data_api — Method

load_cps_data_api(api_key, years; states = STATES)

Load all the CPS data in the given year range.

Arguments

api_key - Request an API key from the Census website

years - The years to pull data for, as a range 2000:2023

Keywords

states - A DataFrame with the state FIPS codes and abbreviations (default is STATES)

Returns

A dictionary with the following keys:

income - A DataFrame with the income data

shares - A DataFrame with the income shares

count - A DataFrame with the count of households

numhh - A DataFrame with the number of households

API Call

There is a split in 2019.

Common	Pre-2019	Post-2019	Description

HWSVAL			wages and salaries

HSEVAL			self-employment (nonfarm)

HFRVAL			self-employment farm

HUCVAL			unemployment compensation

HWCVAL			workers compensation

HSSVAL			social security

HSSIVAL			supplemental security

HPAWVAL			public assistance or welfare

HVETVAL			veterans benefits

HSURVAL			survivors income

HDISVAL			disability

HINTVAL			interest

HDIVVAL			dividends

HRNTVAL			rents

HEDVAL			educational assistance

HCSPVAL			child support

HFINVAL			financial assistance

HOIVAL			other income

HTOTVAL			total household income

GESTFIPS		state fips

A_EXPRRP		expanded relationship code

H_HHTYPE		type of household interview

PPPOS		person identifier

MARSUPWT		asec supplement final weight

HDSTVAL	retirement distributions

HPENVAL	pension income

HANNVAL	annuities

windc_household_data.magic_data_cex — Method

magic_data_cex(output_directory)

Copy the files from the magic_data directory to the output directory. This copies two files, national_income_elasticities_CEX_2013_2017.csv and windc_pce_map.csv.

windc_household_data.magic_data_cps — Method

magic_data_cps(output_directory)

Copy the files from the magic_data directory to the output directory. This copies two files, labor_tax_rates.csv and capital_tax_rates.csv which are created in the SAGE model.

windc_household_data.public_health_benefits — Method

public_health_benefits(cms::DataFrame, acs::DataFrame)

Calculate the public health benefits from the CMS and ACS data.

Arguments

cms::DataFrame: The CMS data the result of get_cms_data or extrapolate_cms_data.

acs::DataFrame: The ACS data the result of get_census_health_data.

Returns

A DataFrame with the following columns:

:state::String: The state name.

:type::String: The income category.

:year::Int: The year of the data.

:income::String: The income category.

:medicare::Float64: The medicare spending.

:medicaid::Float64: The medicaid spending.

Methodology

Group the ACS data by state, income category, and year to compute value shares.

Left join the CMS data to the ACS data.

Calculate the medicare and medicaid spending by multiplying the value shares by the medicare and medicaid spending and dividing by 1000.

windc_household_data.read_commuting_data — Method

read_commuting_data(file_path; states = STATES)

Read the commuting data from the given file path. Download the commuting data with the fetch_commuting_data function.

windc_household_data.save_acs_commuting_data — Method

save_acs_commuting_data(acs_commuting::DataFrame, output_directory)

Save the ACS commuting data to the output directory.

windc_household_data.save_cps_data — Method

save_cps_data(census_data, bea_data, output_directory)

Save the CPS and BEA data to the output directory.

Arguments

census_data - A DataFrame with the CPS data

bea_data - A DataFrame with the BEA data

output_directory - The directory to save the data

windc_household_data.save_public_health_benefits_data — Method

save_public_health_benefits_data(health::DataFrame, output_directory)

Save the public health benefits data to the output directory.